requirements.txt
aws-cdk-lib==2.118.0
constructs>=10.0.0,<11.0.0
openai
boto3>=1.28.57
botocore>=1.31.5
Pillow
Colr
requests
aws_cdk.aws_batch_alpha>=2.72.0a
aws-cdk.aws-amplify-alpha
aws-cdk.aws-apigatewayv2-alpha
aws-cdk.aws-apigatewayv2-integrations-alpha
aws-cdk.aws-apigatewayv2-authorizers-alpha
aws_cdk.aws_cognito_identitypool_alpha
